% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert2h5se}
\alias{convert2h5se}
\title{convert2h5se}
\usage{
convert2h5se(input, outputPath, compression = 1L)
}
\arguments{
\item{input}{A SummarizedExperiment object, or the path to a saved 
serialization of such an object (i.e. rds file).}

\item{outputPath}{The output folder and prefix for the h5SE.}

\item{compression}{The compression level to use.}
}
\value{
An invisible H5 SummarizedExperiment object.
}
\description{
convert2h5se
}
