% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dround}
\alias{dround}
\title{dround}
\usage{
dround(x, digits = 3, roundGreaterThan1 = TRUE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{digits}{The number of digits to keep}

\item{roundGreaterThan1}{Whether to trim also numbers greater than 1 (default TRUE)}
}
\value{
A numeric vector of the same length as `x`
}
\description{
Trim to a certain number of digits (equivalent to `format(...,digits=digits)`, except that the output is numeric)-
}
\examples{
dround( c(0.00002345, 554356, 12.56) )
}
