% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{grepGene}
\alias{grepGene}
\title{grepGene}
\usage{
grepGene(x, g, ignore.case = TRUE)
}
\arguments{
\item{x}{A vector of name composites (e.g. ensemblID.symbol), or a
SummarizedExperiment object with such row names}

\item{g}{A set of genes names to select}

\item{ignore.case}{Logical; whether to ignore case when matchiing}
}
\value{
The subsetted elements of `x` matching `g`
}
\description{
Finds genes names in a vector or Summarized Experiment object
}
\examples{
x <- c("ENSG00000000003.TSPAN6", "ENSG00000000005.TNMD",
       "ENSG00000000419.DPM1")
grepGene(x, c("TSPAN6", "TNMD"))
}
