% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tinySEV}
\alias{tinySEV}
\title{A shiny SummarizedExperiment viewer}
\description{
A shiny SummarizedExperiment viewer
}
\examples{
ui <- tinySEV.iu()
server <- tinySEV.server(objects=c(SE1="path/to/SE1.rds",
                                   SE2="path/to/SE2.rds"))
}
