% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{tinySEV.server}
\alias{tinySEV.server}
\title{tinySEV.server}
\usage{
tinySEV.server(
  objects = NULL,
  uploadMaxSize = 50 * 1024^2,
  maxPlot = 500,
  genelists = list(),
  genelistTab = length(genelists) > 0
)
}
\arguments{
\item{objects}{A named list of (paths to)
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} objects.}

\item{uploadMaxSize}{The maximum upload size. Set to zero to disable upload.}

\item{maxPlot}{The maximum number of features to allow for plotting heatmaps.}

\item{genelists}{An optional named list of genes/features which will be 
flagged in the gene tab.}
}
\value{
A shiny server function.
}
\description{
tinySEV.server
}
