% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tinySEV.ui}
\alias{tinySEV.ui}
\title{tinySEV.ui}
\usage{
tinySEV.ui(
  title = "tinySEV",
  waiterContent = NULL,
  about = NULL,
  skin = "blue",
  hasLogin = FALSE
)
}
\arguments{
\item{title}{The title of the app (displayed in the header)}

\item{waiterContent}{Optional content of the loading mask; should be a
`tagList`, NULL to use default, or FALSE to disable the waiter}

\item{about}{Optional content of the introduction page (NULL to disable
intro page)}

\item{skin}{The dashboard skin color, passed to 
\code{\link[shinydashboard]{dashboardPage}}.}
}
\value{
a shiny UI
}
\description{
tinySEV.ui
}
