% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{tinySEV.server}
\alias{tinySEV.server}
\title{tinySEV.server}
\usage{
tinySEV.server(
  objects = NULL,
  uploadMaxSize = 50 * 1024^2,
  maxPlot = 500,
  feature.lists = list(),
  filelist = list(),
  logins = NULL,
  feature.listsTab = TRUE
)
}
\arguments{
\item{objects}{A named list of (paths to)
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} objects.}

\item{uploadMaxSize}{The maximum upload size. Set to zero to disable upload.}

\item{maxPlot}{The maximum number of features to allow for plotting heatmaps.}

\item{feature.lists}{An optional named list of genes/features which will be 
flagged in the gene tab.}

\item{filelist}{A named list of downloadable files optionally associated with
elements of `objects`, or a folder where to find these files}

\item{logins}{An optional dataframe containing possible logins. Must contain 
the columns "user" and "password_hash" (sodium-encoded). Not providing the
argument disables login.}

\item{feature.listsTab}{Logical, whether to show the feature list tab}
}
\value{
A shiny server function.
}
\description{
tinySEV.server
}
